+++ migrations/Version20180905190737.php
@@ -24,28 +24,36 @@ final class Version20180905190737 extends AbstractMigration
 {
     public function up(Schema $schema): void
     {
-        $this->addSql('ALTER TABLE kimai2_activities ADD COLUMN fixed_rate NUMERIC(10, 2) DEFAULT NULL');
-        $this->addSql('ALTER TABLE kimai2_activities ADD COLUMN hourly_rate NUMERIC(10, 2) DEFAULT NULL');
+        if ($this->isPlatformMySQL()) {
+            $this->addSql('ALTER TABLE kimai2_activities ADD COLUMN fixed_rate NUMERIC(10, 2) DEFAULT NULL');
+            $this->addSql('ALTER TABLE kimai2_activities ADD COLUMN hourly_rate NUMERIC(10, 2) DEFAULT NULL');
 
-        $this->addSql('ALTER TABLE kimai2_projects ADD COLUMN fixed_rate NUMERIC(10, 2) DEFAULT NULL');
-        $this->addSql('ALTER TABLE kimai2_projects ADD COLUMN hourly_rate NUMERIC(10, 2) DEFAULT NULL');
+            $this->addSql('ALTER TABLE kimai2_projects ADD COLUMN fixed_rate NUMERIC(10, 2) DEFAULT NULL');
+            $this->addSql('ALTER TABLE kimai2_projects ADD COLUMN hourly_rate NUMERIC(10, 2) DEFAULT NULL');
 
-        $this->addSql('ALTER TABLE kimai2_customers ADD COLUMN fixed_rate NUMERIC(10, 2) DEFAULT NULL');
-        $this->addSql('ALTER TABLE kimai2_customers ADD COLUMN hourly_rate NUMERIC(10, 2) DEFAULT NULL');
+            $this->addSql('ALTER TABLE kimai2_customers ADD COLUMN fixed_rate NUMERIC(10, 2) DEFAULT NULL');
+            $this->addSql('ALTER TABLE kimai2_customers ADD COLUMN hourly_rate NUMERIC(10, 2) DEFAULT NULL');
+        } else {
+            $this->preventEmptyMigrationWarning();
+        }
     }
 
     public function down(Schema $schema): void
     {
-        $customer = $schema->getTable('kimai2_customers');
-        $customer->dropColumn('hourly_rate');
-        $customer->dropColumn('fixed_rate');
+        if ($this->isPlatformMySQL()) {
+            $customer = $schema->getTable('kimai2_customers');
+            $customer->dropColumn('hourly_rate');
+            $customer->dropColumn('fixed_rate');
 
-        $project = $schema->getTable('kimai2_projects');
-        $project->dropColumn('hourly_rate');
-        $project->dropColumn('fixed_rate');
+            $project = $schema->getTable('kimai2_projects');
+            $project->dropColumn('hourly_rate');
+            $project->dropColumn('fixed_rate');
 
-        $activity = $schema->getTable('kimai2_activities');
-        $activity->dropColumn('hourly_rate');
-        $activity->dropColumn('fixed_rate');
+            $activity = $schema->getTable('kimai2_activities');
+            $activity->dropColumn('hourly_rate');
+            $activity->dropColumn('fixed_rate');
+        } else {
+            $this->preventEmptyMigrationWarning();
+        }
     }
 }
