+++ migrations/Version20190510205245.php
@@ -23,23 +23,31 @@ class Version20190510205245 extends AbstractMigration
 {
     public function up(Schema $schema): void
     {
-        $timesheetTags = $schema->createTable('kimai2_timesheet_tags');
-        $timesheetTags->addColumn('timesheet_id', 'integer', ['length' => 11, 'notnull' => true]);
-        $timesheetTags->addColumn('tag_id', 'integer', ['length' => 11, 'notnull' => true]);
-        $timesheetTags->addIndex(['timesheet_id'], 'IDX_E3284EFEABDD46BE');
-        $timesheetTags->addIndex(['tag_id'], 'IDX_E3284EFEBAD26311');
-        $timesheetTags->setPrimaryKey(['timesheet_id', 'tag_id']);
+        if ($this->isPlatformMySQL()) {
+            $timesheetTags = $schema->createTable('kimai2_timesheet_tags');
+            $timesheetTags->addColumn('timesheet_id', 'integer', ['length' => 11, 'notnull' => true]);
+            $timesheetTags->addColumn('tag_id', 'integer', ['length' => 11, 'notnull' => true]);
+            $timesheetTags->addIndex(['timesheet_id'], 'IDX_E3284EFEABDD46BE');
+            $timesheetTags->addIndex(['tag_id'], 'IDX_E3284EFEBAD26311');
+            $timesheetTags->setPrimaryKey(['timesheet_id', 'tag_id']);
 
-        $tags = $schema->createTable('kimai2_tags');
-        $tags->addColumn('id', 'integer', ['length' => 11, 'autoincrement' => true, 'notnull' => true]);
-        $tags->addColumn('name', 'string', ['length' => 100, 'notnull' => true]);
-        $tags->addUniqueIndex(['name'], 'UNIQ_27CAF54C5E237E06');
-        $tags->setPrimaryKey(['id']);
+            $tags = $schema->createTable('kimai2_tags');
+            $tags->addColumn('id', 'integer', ['length' => 11, 'autoincrement' => true, 'notnull' => true]);
+            $tags->addColumn('name', 'string', ['length' => 100, 'notnull' => true]);
+            $tags->addUniqueIndex(['name'], 'UNIQ_27CAF54C5E237E06');
+            $tags->setPrimaryKey(['id']);
+        } else {
+            $this->preventEmptyMigrationWarning();
+        }
     }
 
     public function down(Schema $schema): void
     {
-        $schema->dropTable('kimai2_timesheet_tags');
-        $schema->dropTable('kimai2_tags');
+        if ($this->isPlatformMySQL()) {
+            $schema->dropTable('kimai2_timesheet_tags');
+            $schema->dropTable('kimai2_tags');
+        } else {
+            $this->preventEmptyMigrationWarning();
+        }
     }
 }
